replacing VaxrankCoreLogic class with module full of plain functions with explicit signatures
